refactor: move WorkflowParser and WorkflowYaml to workflow package
